new routes for login and register
