--- pages/authentication/login/index.tsx
@@ -1,42 +1,0 @@
-import { Box, Button, Grid, TextField } from '@material-ui/core'
-import React from 'react'
-import classes from './style.module.scss'
-
-const Login = (): JSX.Element => {
-    return (
-        <Box boxShadow={3}>
-            <form className={classes.form} noValidate autoComplete="off">
-                <Grid container spacing={4}>
-                    <Grid item xs={12}>
-                        <h1>{'CONNEXION'}</h1>
-                    </Grid>
-                    <Grid item xs={12}>
-                        <TextField id="email-input" fullWidth label="Email" />
-                    </Grid>
-                    <Grid item xs={12}>
-                        <TextField
-                            id="password-input"
-                            fullWidth
-                            label="Mot de passe"
-                            type="password"
-                        />
-                    </Grid>
-                    <Grid item xs={12}>
-                        <Button
-                            classes={{ root: classes.button }}
-                            variant="contained"
-                            color="primary"
-                            onClick={() => {
-                                window.location.pathname = '/'
-                            }}
-                        >
-                            {'Se connecter'}
-                        </Button>
-                    </Grid>
-                </Grid>
-            </form>
-        </Box>
-    )
-}
-
-export default Login
