+++ pages/register/style.module.scss
@@ -15,3 +15,8 @@
 .button {
     background-color: $secondary-dark !important;
 }
+
+.container {
+    display: flex;
+    justify-content: center;
+}
